\name{fishercpp}
\alias{fishercpp}
\title{Fisher information matrix of the log partial likelihood of a
  multistate model.}
\description{This function provides a fast implementation for the calculation of the Fisher information matrix
  needed for the estimation of fusion lasso penalized multi-state models in a piece-wise exponential framework.}
\usage{
fishercpp(Xcpp, mucpp)
}
\arguments{
  \item{Xcpp}{...}
  \item{mucpp}{...}
}
\details{...}
\value{...}
\author{Holger Reulen}
\examples{
\dontrun{fishercpp(Xcpp, mucpp)}
}